refactor: separate utility functions and business logic

+++ src/shield.js
@@ -1,7 +1,8 @@
 import fs from 'fs';
+import { isAbsolute } from 'path';
 import { execSync } from 'child_process';
-import { config } from './config';
 import { isValidApp } from './block';
+import { config, editConfig } from './config';
 import { generatePassword, sha256 } from './utils';
 
 export const isValidPassword = (password) => {
@@ -10,29 +11,25 @@ export const isValidPassword = (password) => {
     return sha256sum === config.passwordHash;
 };
 
-export const enableShieldMode = (password = generatePassword()) => {
+export const enableShieldMode = async (password = generatePassword()) => {
     const passwordHash = sha256(password);
     console.log(`Your password is: ${password}`);
 
-    config.password = password;
-    config.passwordHash = passwordHash;
-    config.shield = true;
+    await editConfig({ ...config, password, passwordHash, shield: true });
 };
 
-export const disableShieldMode = (password) => {
-    if (isValidPassword(password)) {
-        config.shield = false;
-        delete config.passwordHash;
-    }
+export const disableShieldMode = async (password) => {
+    await editConfig({ ...config, password, shield: false });
 };
 
 export const blockRoot = () => {
     if (process.env.NODE_ENV === 'test') return;
+
     execSync('usermod -s /usr/sbin/nologin root');
     fs.writeFileSync('/etc/sudoers.d/ulysse', `${process.env.SUDO_USER} ALL=(ALL) !ALL`, 'utf8');
 
     for (const w of config.whitelist) {
-        if (isValidApp(w.name)) {
+        if (isValidApp(w.name) && isAbsolute(w.name)) {
             fs.appendFileSync('/etc/sudoers.d/ulysse', `\n${process.env.SUDO_USER} ALL=(ALL) ${w.name}`, 'utf8');
         }
     }
@@ -41,6 +38,8 @@ export const blockRoot = () => {
 };
 
 export const unblockRoot = () => {
+    if (process.env.NODE_ENV === 'test') return;
+
     execSync('usermod -s /bin/bash root');
     if (fs.existsSync('/etc/sudoers.d/ulysse')) {
         fs.unlinkSync('/etc/sudoers.d/ulysse');
